Switch to built-in res.cookie() for CSRF token
Doing this because I just realized that Express has its own built-in
cookie functionality and doesn't need an extra npm package to do basic
cookie stuff. Cookie signing is made a bit easier by middleware, and
indeed I am using the 'cookie-session' npm package to manage the signed
session cookie.

Then I decided it's not even necessary to sign the CSRF token cookie.

So I ditched the 'cookies' npm package.

Ok, but the real reason I wanted to ditch 'cookies' is because I was
getting a stubborn type error while trying to use it as middleware. To
summarize, 'cookies' says it supports this usage:

    import express from 'express'
    import Cookies from 'cookies'

    const app = express()
    app.use(Cookies.express(['mykey']))

This middleware adds the property 'cookies' on Express's Request and
Response objects. For instance:

    app.get('/', (req, res) => {
        console.log(req.cookies.get('a'))
        res.cookies.set('b', 'c')
    })

And, this actually works! (at runtime!)

However, at the type checking stage, the Typescript compiler complains
that there is no property 'cookies' on 'res', or 'req'. Frustratingly,
installing the '@types/cookies' npm package does not fix this.

There's probably an opportunity to contribute to '@types/cookies' here,
but... somehow that feels futile. The larger issue is a systemic one.